Handle @Relation annotation with non-list type of more than 999 rows.
When fetching one-to-one relationships where there are more than 999
parent items we need to batch query the children. We already had the
mechanism to do this for one-to-many and many-to-many but for one-to-one
the same strategy failed because the batching code relied in the value
of the relating map to be a collection, a reference that can be passed
around and filled separately. Meanwhile for one-to-one we have to
copy the batch map into the original relationship map.

Bug: 143105450
Test: PojoWithRelationTest, BareRelationDatabaseTest
Change-Id: I00fac5076d4243e973d0e9e2801fef659d518e05
(cherry picked from commit 8331d96ce8ab2b68f391211ffaecb8db32be5a5c)