copy new native libs if they are loaded in another classloader
This is an attemp to fix b/146061836 by copying sqlite native
libraries into another temp folder so that we can load them
as if they are different classes.

I cannot reproduce the issue locally (the stress test that was
created for the previous issue passes now) but this is worth
a try. ¯\_(ツ)_/¯

Bug: 146061836
Test: n/a
Change-Id: I2492a48f866386702e19478114ebee47d3f08760
(cherry picked from commit d8a14e6b32aedbbf9c53105ebc40fad702a70277)