Tie transaction job to the calling context
This CL fixes a bug in room suspend transactions where we would deadlock
the transactions if the calling coroutine gets cancelled before we can
even start the transaction. It would leave the control job in active
state with no other callback to cancel it (to unlock the runner).

With this change, we automatically cancel it if the calling context
has a job. Note that it is not added as a child job since doing so would
cancel the calling context when job is cancelled. We want that to be
just a one way cancelation.

Bug: 148181325
Test: SuspendingTransactionCancelationTest
Change-Id: I18a76128d325f099b9677af1bb5d35cedf43e3d5
(cherry picked from commit c89f069d9d7de77f8f0961f15f318494687cdac2)