Properly inherit type argument types in JvmWildcardResolution
This CL fixes two bugs in JvmWildcardResolution, one of which was hidden
by the other one.

Previously, when there is a different declaration type, we've always
obtained the variance from the type parameter declaration. This was
wrong because the inherited type might've already declared an explict
variance. Parameter type variance is only relevant if the declaration
type is a TypeParameterType (e.g. T in List<T>).

Fixing that one revealed another bug where, while resolving wildcards
for synthetic setter parameters, we were using the field's type instead
of the parameter type from setter. They are different in JVM since the
parameter will have jvm wildcards applied whereas field does not
participate in wildcards. To keep that part clean, I've moved the
synthetic accessors to the field rather than computing them in
KspTypeElement.

Last but not least, I've also made another consistency change on how we
handle continuation. When a suspend function is converted to JVM, it
receives an additional Continuation parameter which has the return type
as a type argument. Continuation's type argument is declared as `in` so
we always applied it while creating the type. But it is not really
correct because, from Kotlin's perspective, it is invariant. To be more
consistent, I've changed it to be invariant as well and let java code
generation add the variance.

Bug: 217210973
Test: KspTypeNamesGoldenTest
Change-Id: I970a79505452ca10da9ba7151323aed699ba7f2f
(cherry picked from commit 2250d6733b37f8fc0a14baba8edef30f74813a33)
Merged-In:I970a79505452ca10da9ba7151323aed699ba7f2f