Handle reentrant case in Room's acquireTransactionThread()
During a suspending transaction Room uses a thread from the transaction executor, starts an event loop in it and dispatches suspending database operations to it so they are all encapsulated within the transaction coroutine. It is usually expected that the transaction thread is different from the one starting the transaction, but in some cases they are the same. To handle such reentrant cases the `withTransaction()` has been refactored to no longer rely joining on a control job and instead it will execute the suspending transaction block from within the `runBlocking` in the transaction thread.

Executing the suspending transaction block in the `runBlocking` prevents a deadlock situation due to the continuation bouncing between the caller to the transaction thread and back. The deadlock specifically happened when Room started the event loop via `runBlocking`, resumed the `acquireTransactionThread()` coroutine but also joined the `runBlocking` coroutine with the control job. These set of events would cause a deadlock when both the acquire thread and transaction thread are the same because `acquireTransactionThread()` is unable to continue as it is waiting for `runBlocking` to finish which in turns waits for the control job to be done. Meanwhile the job never completes because the database transaction does not finish, i.e. `TransactionElement.release` is never called. By eliminating the need to wait on a control job but still support nested transactions via the transaction context, Room is able to avoid the deadlock situation.

This behaviour was mostly observed in tests where the transaction executor is usually an instant executor, but it could happen in an actual application if Room is configured to share a thread pool executor as the transaction executor and a background coroutine dispatcher and the thread pool is sufficiently busy that only one thread is left for both starting the transaction (to acquire the transaction thread) and the transaction thread itself.

Bug: 251503005
Test: SuspendingQueryTest
Change-Id: I319d977a080b89880d5cf68d139c69a8ba096107
(cherry picked from commit on android-review.googlesource.com host: 7d8a5b162ba1a8814da51ba950b8555ca94895b3)
Merged-In: I319d977a080b89880d5cf68d139c69a8ba096107
(cherry picked from commit on android-review.googlesource.com host: a613c4f3691d3cb9ae3b9f869778178ab76be230)
Merged-In: I319d977a080b89880d5cf68d139c69a8ba096107