Use an isOpenInternal check when checking if DB is open.
For compatibility reasons the public API isOpen will return true when auto-close is enabled and the DB is closed, but sometimes we need to check if the actual connection is open or not to do a thread jump or avoid a potential disk IO call, such as the one done in CoroutineRoom. To solve this, this change adds an isOpenInternal that is restricted and checks if the actual connection is open or not.

Bug: 262255165
Test: SuspendingQueryTest.allBookSuspend_autoClose
Change-Id: I5f39308a80eab2242be7764dd3854ebfd774c195
(cherry picked from commit on android-review.googlesource.com host: 1988e1edfa5b73a875ac7a8eb97dde3b0fd3e610)
Merged-In: I5f39308a80eab2242be7764dd3854ebfd774c195