Reimplement suspending pool with Semaphore
This changes the implementation of Room's connection pool to use a Semaphore instead of a Channel. It makes the implementation simpler and removes the possibility of connections getting 'lost' due the the various branches needed to handle when a channel send / receive fails.

The Channel was also used as a 'queue' instead a circular array is used to push and pop connections acquired and released into the pool. The semaphore still controls the amount of connections being locked and unlocked and is the suspending API that waits for a connection to be free.

Bug: 322386871
Bug: 380088809
Test: BaseConnectionPoolTest
(cherry picked from https://android-review.googlesource.com/q/commit:dc4715197f7869c87af3025acba91a0c8a57db4d)
Merged-In: I9b7f2d44b511eb3e3d8cdaf8cb8905bb6203f8c5
Change-Id: I9b7f2d44b511eb3e3d8cdaf8cb8905bb6203f8c5