Log connection timeout instead of throwing
The original idea of using withTimeout while acquiring a connection was to prevent deadlocks due to a coroutine never releasing a connection. This change updates the pool to just log the busy exception that would have been thrown in order to help developers if they notice a deadlock but not crash the app.

The change from throwing to logging is done to support iOS looper suspension. Kotlin Coroutines rely on iOS Foundation timers and those themselves rely on Run Loops. When an iOS app is backgrounded the process is suspended along with the loopers, when the app is resumed all events that would have triggered while the app was in the background are immediately fired, causing Room to throw the timeout exception if there was a coroutine waiting for a connection before the app was suspended. The timeout in Room is meant to represent time waiting for a connection while the process is active, not clock time. In order to support active only timeouts we would need listeners for foreground / background states changes but those are not accessible at the layer where Room operates, therefore instead of throwing Room will just log and even though it might log false-positives when the app is resumed it will not crash the app and the log will still be useful if there is a real deadlock.

Bug: 380088809
Bug: 422448815
Test: BundledSQLiteConnectionPoolTest
(cherry picked from https://android-review.googlesource.com/q/commit:a1ea1d8a577d858b1d2bdb301cd28a1efcb008aa)
(cherry picked from https://android-review.googlesource.com/q/commit:c9b0c6beb5dd96c8ad198a226c1763bd577580cf)
Merged-In: I78a29565acc98d6ff8d2074efe965b0fa570aef2
Change-Id: I78a29565acc98d6ff8d2074efe965b0fa570aef2