Resolve Room incorrectly triggering `fallbackToDestructiveMigration` when a migration path exists.
In `PrePackagedCopyOpenHelper` when the database file verification is being performed, the check for whether or not a migration is "required" (otherwise falling back to a destructive migration) only checked for:
[1] Whether the provided "from" and "to" versions resulted in a downgrade, in which case a migration would not be required.
[2] There is a general requirement for migrations but the versions provided have an exception for not requiring a migration "from" that version.

Neither of these checks considered whether or not there already exists any migration paths present in the migration container of the database configuration. This results in the behavior reported in the bug, where a `fallbackToDestructiveMigration` overrides all existing migrations in the migrationContainer of the database configuration. Checking if the provided "from" and "to" versions have a migration path *before* checking for scenario [1] and [2] sufficiently avoids incorrectly falling back to a destructive migration in the case where there is an automigration or a manual migration provided.

Bug: 432634197
Test: Tested with sample project from bug using ./gradlew publishToMavenLocal
(cherry picked from https://android-review.googlesource.com/q/commit:f902012bb7bd8728a4c80bc15b32011e3973f735)
Merged-In: I7cf209d3e2632e163a81ed6137012d7704aad6c4
Change-Id: I7cf209d3e2632e163a81ed6137012d7704aad6c4