Combine sync trigger critical section
This change partially reverts Ic9a3c7c3470578d08467841e552845b4787ca8af and attempts to tackle the syncing triggers race again.

In the previous attempt before a connection is used the `onSync` function is called and in it a lock is grabbed and released before the action to actually create the TRIGGERs, once the action was done the lock was grabbed again and various cases of things that could happen due to releasing the lock where being handled by checking the 'needSync' state again and using a 'version'. The big mistake was that creating and dropping triggers operations where being committed in the transaction even though the version of the sync didn't match and the observed state of the tables was not being updated, creating an inconsistency between the installed TRIGGERs and the 'tableObservedState'.

In this attempt, the optimization to not grab a connection if no sync is needed is dropped and things are reverted to always grabbing a connection when performing a sync, however, the optimization of not starting a transaction is still maintained, avoiding re-introducing the transaction lock that has been problematic in the past. This attempt simplifies things by including the action (creating and dropping TRIGGERs) within the ObservedTableStates lock such that those that add and remove observers that result in pending syncs (waiting for the sync) do actually wait and not go ahead with their queries.

Bug: 432365736
Test: FlowQueryTest
(cherry picked from https://android-review.googlesource.com/q/commit:e1b90d98c6670350c6d2b5e477ebf3e3baad4a32)
Merged-In: I72380efaf85a35f0037fa1265f5e67d4cb0d6dc9
Change-Id: I72380efaf85a35f0037fa1265f5e67d4cb0d6dc9