Resolve circular lock between AutoCloser and InvalidationTracker
This change partially reverts I72380efaf85a35f0037fa1265f5e67d4cb0d6dc9 and attempts to tackle the syncing triggers race yet again while also avoiding a deadlock introduced by the mentioned CL.

It was a bad idea to lock around a suspend function (the onSync() action) that will also do a database operation, that is a bad practice in general.

This change goes back to a similar strategy done in Ic9a3c7c3470578d08467841e552845b4787ca8af without the version complexities.

There are two parts, fixing the deadlock and not break highly concurrent invalidation.

The deadlock was caused by a circular lock dependency between the AutoCloser and the InvalidationTracker. When a database is being re-opened after an auto-close it has to reconfigure the database with the InvalidationTracker and calls forceNeedSync() which attempts to grab the ObservedTableStates lock while also holding the AutoCloser lock. Meanwhile if a concurrent Flow is created it will attempt to sync triggers, and that grabs the ObservedTableStates lock first before then trying to use the database to add triggers which will need the AutoCloser lock, already held by the reconfiguring of the database. The solution is to simply not hold the ObservedTableStates lock during the database transaction that add / removes triggers, avoiding the circular dependency.

Removing the dependency then re-introduces the issues with reactive return types missing invalidation because a new Flow will no longer wait for syncTriggers() to complete, the 'action'`' of onSync() is no longer within the ObservedTableStates lock, on-purpose to avoid the deadlock. Therefore in order to serialize trigger syncs and guaranteed that newly created flows wait for trigger installation a small lock and flag are used. The flag is flipped ON at the start of onSync() and OFF at the end. It signals onObserverAdded() and onObserverRemoved() that a sync should be done even if needsSync is false. This makes it such that syncTrigger() is attempted which in turn can guaranteed trigger installation or serve as a waiting mechanism thanks to the second item, the small lock. The lock is focused and only covers onSync() specifically. It is used solely for serializing sync attempts. Note that within a block of ObservedTableStates's lock, the newly added syncTriggerLock is never requested, so no new deadlock is introduced.

Bug: 446643789
Test: AutoClosingDatabaseTest and FlowQueryTest
Cherrypick-From: https://android-review.googlesource.com/q/commit:4a9a26481ac25a468243ed839017877b0e33d144
Merged-In: I5d47a402f278da7d96a4c50e86c2dbc3272287f9
Change-Id: I5d47a402f278da7d96a4c50e86c2dbc3272287f9