Avoid transaction thread hop in Room's SupportSQLite Wrapper
This change updates the thread confined transaction logic in the wrapper to use Dispatchers.Unconfined, where a connection is acquired and hold on a launched coroutine for the duration of the transaction. This makes it so that the connection holding coroutine is performed in the same thread as the caller thus making the wrapper perform follow-up database operation for the transaction in the same thread, blocking as it is done for non-transaction calls.

It is OK to use Dispatchers.Unconfined for this case even though it causes the coroutine to have no thread affinity, which is sort of the oppositive of what we are trying to do, 'confine' the connection to a thread. The reason it is OK is because there are no other dispatchers changes done by Room once useConnection() is called and so there is no concern of the coroutine resuming in a different thread. Moreover, since the wrapper uses runBlocking, when dispatching again into the transaction coroutine, it will use the Unconfined dispatcher that is in the transaction context acquired earlier when the transaction started, causing the database operation that is part of the transaction to be executed in the same thread, accomplishing the same blocking behaviour as Android's SQLite APIs.

Bug: 441117897
Test: SupportSQLiteDatabaseWrapperTest
Cherrypick-From: https://android-review.googlesource.com/q/commit:42acadf200217a00816339ccceb07be785721d12
Merged-In: I00dd89549e09281c83718a95e9bd7dcb18ebbd3d
Change-Id: I00dd89549e09281c83718a95e9bd7dcb18ebbd3d